+++ components/Calendar/Calendar.tsx
@@ -11,19 +11,19 @@ function Calendar() {
   moment.updateLocale("en", { week: { dow: 1 } });
 
   const [today, setToday] = useState(moment());
-
   const startDay = today.clone().startOf("month").startOf("week");
 
   const prevHandler = () =>
     setToday((prev: any) => prev.clone().subtract(1, "month"));
+
   const todayHandler = () => setToday(moment());
+
   const nextHandler = () =>
     setToday((prev: any) => prev.clone().add(1, "month"));
 
   return (
     <div className={classes.container}>
       <div className={classes.main}>
-        <Title />
         <HeaderCalendar
           prevHandler={prevHandler}
           todayHandler={todayHandler}
