+++ styles/colors.scss
@@ -22,5 +22,5 @@ $title_background: #2a2820;
 
 $main_border_calendar: #464648;
 
-$black_01: #1e1f21;
-$black_02: #484848;
+$black_01: #ffffff;
+$black_02: #eeeeee;
